feat: clear case id page and separate metrics and table component
