+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@nivo/bar": "^0.87.0",
     "@nivo/line": "^0.87.0",
+    "@nivo/pie": "^0.87.0",
     "@radix-ui/react-collapsible": "^1.0.3",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-navigation-menu": "^1.1.4",
@@ -19,6 +20,7 @@
     "chart.js": "^4.4.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "lodash": "^4.17.21",
     "lucide-react": "^0.378.0",
     "next": "14.2.3",
     "react": "^18",
@@ -29,6 +31,7 @@
     "zustand": "^4.5.2"
   },
   "devDependencies": {
+    "@types/lodash": "^4.17.4",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
